requirements.txt
python-qbittorrent
scrapy
clint
